fix(remote): follow-up fixes — deduplication, safety, daemon wiring (swamp-club#969)
Correctness:
- Add AbortSignal.timeout(60s) to enrollment verify dispatch — prevents
  indefinite hangs if the probe worker stalls without disconnecting.
- Scope probeMarker injection to swamp/fleet-probe model type only —
  prevents blind merge of dispatch-level field into arbitrary models.
- Guard #runEnrollmentVerify against disconnected workers — check
  entry.channel === null before persisting status after probe completes.
- Move entry.verifyFailureReason mutation inside #recordTransition so
  in-memory and durable state stay consistent.

Deduplication:
- Extract parseLabels to src/cli/commands/worker_shared.ts — was
  duplicated identically in worker_connect.ts and worker_verify.ts.
- Move WorkerProbeResult and WorkerVerifyData to src/serve/protocol.ts —
  was duplicated in worker_verify.ts and admin_handlers.ts. Fixes the
  presentation → CLI import inversion (worker_output.ts now imports
  from serve/protocol.ts).
- Extract dispatchFleetProbe helper to src/serve/fleet_probe_dispatch.ts —
  deduplicates the dispatch+read+parse logic from serve.ts and
  admin_handlers.ts into a single shared function.

UX improvements:
- Add hint in renderWorkerList when unverified workers are present:
  "Some workers are unverified — run 'swamp worker verify' to diagnose."
- Distinguish fail vs error in renderWorkerVerify summary:
  "1 failed, 1 unreachable" instead of "2 failed verification."

Daemon + env var:
- Wire --verify-on-enroll into daemonEnableCommand and
  collectServeExtraArgs so daemon deployments can enable it.
- Add SWAMP_VERIFY_ON_ENROLL env var fallback for the flag.
